+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # 2.x branch
 ## 2.21 branch
 ### 2.21.4
-* added `li()` method for `BootstrapHtmlHelper`.
+* added `li()` and `meta()` methods for `BootstrapHtmlHelper`.
 
 ### 2.21.3
 * added `checkbox()`, `radio()` and `select()` methods for `BootstrapFormHelper`;